+++ b2b-app/imports/api/forms/engine.js
@@ -1,11 +1,12 @@
-import { createTypeErrorMsg } from 'pdf-lib'
+import { getError } from './engine-errors'
 
 const debug = require('debug')('app:forms:engine')
 
-let survey = { steps: [] }
+let survey = { sections: [] }
 let currentStep
 let currentQ
 let current
+let errs
 
 const slugify = (text) => {
   if (!text || typeof text !== 'string') return 'no-slug'
@@ -23,8 +24,8 @@ const addQ = (survey, prompt) => {
 }
 
 const addStep = (survey, title) => {
-  currentStep = { name: title, questions: [], id: slugify(title) }
-  survey.steps.push(currentStep)
+  currentStep = { title, questions: [], id: slugify(title) }
+  survey.sections.push(currentStep)
   return currentStep
 }
 
@@ -35,8 +36,10 @@ const addGrid = (survey, title) => {
 }
 
 const addAnswer = (survey, text) => {
-  currentA = { name: text, id: slugify(text), type: 'text' }
-  currentQ.answers.push(currentA)
+  currentA = { title: text, id: slugify(text), type: 'text' }
+  if (!currentQ) {
+    return { errCode: 'e-no-ans' }
+  } else currentQ.answers.push(currentA)
   return currentA
 }
 
@@ -52,16 +55,17 @@ objects.forEach((o) => {
 })
 
 export const parse = (source) => {
+  currentStep = currentQ = current = null
+  errs = []
   if (typeof source !== 'string') throw new Error('Parameter to parse must be a string')
   try {
     survey = {
-      steps: [],
+      sections: [],
       name: 'Sample Survey',
       slug: 'sample',
       version: '1',
       active: true,
     }
-    const result = { status: 'failed', errs: [] }
     const lines = source.split('\n').map((line) => line.trim())
     lines.forEach((line, ix) => {
       const lineno = ix + 1
@@ -75,7 +79,9 @@ export const parse = (source) => {
           if (m) {
             // debug(`${o.name}: ${m[1]}`)
             if (o.method) {
-              current = o.method(survey, m[1])
+              const res = o.method(survey, m[1])
+              if (res.errCode) errs.push({ lineno, errCode: res.errCode, line })
+              else current = res
             }
             got = true
           }
@@ -85,14 +91,27 @@ export const parse = (source) => {
           if (m) {
             got = true
             const [match, key, value] = m
-            current[key] = value || true
+            if (!current)
+              errs.push({
+                lineno,
+                errCode: 'e-no-obj',
+                line,
+              })
+            else current[key] = value || true
           }
         }
-        if (!got) result.errs.push({ lineno, error: `I could not understand`, line })
+        if (!got) errs.push({ lineno, errCode: 'e-unk', line })
       }
     })
-    if (result.errs.length) return result
-    return { status: 'success', survey }
+    if (errs.length) {
+      return {
+        status: 'failed',
+        message: '',
+        errs: errs.map((err) => ({ error: getError({ code: err.errCode }), ...err })),
+      }
+    }
+
+    return { status: 'success', message: '', survey }
   } catch (e) {
     return { status: 'exception', message: e.message }
   }
